requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
pillow==11.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
requests==2.32.3
tk==0.1.0
typing_extensions==4.12.2
urllib3==2.2.3
